Safari-611.1.10/LayoutTests/js/dom/constructor-with-return-masquerades-expected.txt

This tests the constructor returning masquerades as undefined return masquerades.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result is document.all
PASS successfullyParsed is true

TEST COMPLETE
